refactor and cleanup ginis official board fetching and remove legacy api routes and files
