+++ next/backend/ginis/types.ts
@@ -1,0 +1,6 @@
+export type ParsedOfficialBoardDocument = {
+  id: string
+  title: string
+  createdAt: string
+  content: string
+}
