+++ next/backend/ms-graph/fetchers/msGraphSearch.fetcher.ts
@@ -1,7 +1,7 @@
 import { MSGraphFilteredGroupUser } from '@backend/ms-graph/types'
 import axios from 'axios'
 
-export const getMsGraphSearchQueryKey = (search: string) => ['msGraphUserSearch', search]
+export const getMsGraphSearchQueryKey = (search: string) => ['msGraphSearch', search]
 
 export const msGraphSearchFetcher = async (searchQuery: string) => {
   return axios.get<MSGraphFilteredGroupUser[]>(
