+++ next/backend/tootoot/constants.ts
@@ -1,0 +1,37 @@
+export const cityTootootId = '63dabb7abacf0a031cd19693' //	Mesto Bratislava
+
+export const cityOrganizationsTootootIds = [
+  '63d14613826c200250f25391', //	BKIS
+  '638f33a8b96254062cbbc0ac', //	Divadlo P. O. Hviezdoslava
+  '63ceeb9a408a9511ecea8f5e', //	Galéria mesta Bratislava
+  '63d040e7826c20047431c62d', //	Generálny investor Bratislavy: Titulka
+  '63d042702cd37e0a00044e9e', //	Komunálny podnik Bratislavy
+  '63d044442cd37e0a000548ae', //	MARIANUM - Pohrebníctvo mesta Bratislavy
+  '63d066a22cd37e0a00108eff', //	MESTSKÁ KNIŽNICA V BRATISLAVE
+  '63d0685a826c2004744ce753', //	Mestské lesy v Bratislave
+  '63d068e2826c2004744d1e65', //	MESTSKÝ ÚSTAV OCHRANY PAMIATOK V BRATISLAVE
+  '63d06a9b2cd37e0a0010b2d8', //	Metropolitný inštitút Bratislavy
+  '63d06b912cd37e0a0010c10c', //	Múzeum mesta Bratislava
+  '63d06c9d2cd37e0a0010cf9a', //	STaRZ - Správa telovýchovných a rekreačných zariadení hlavného mesta SR
+  '63d06dc32cd37e0a0010d19c', //	Zoologická záhrada
+  '63ee15be1ff4a0077c9e14ae', //	Bratislava Tourist Board
+  '63ee18a5049c0b04802ae407', //	Archív mesta Bratislava
+  '6407c1d1caa5e70f74c9dc1f', //	OLO Bratislava
+  '63d1574f826c200250fb59f4', //	Mestská časť Petržalka
+  '63d25e04bacf0a0f6c8a1b58', //	Mestská časť Čuňovo
+  '63d25fb7bacf0a0f6c8abf59', //	Mestská časť Devín
+  '63d261fbbacf0a0f6c8c05c3', //	Mestská časť Devínska Nová Ves
+  '63d263ab965fb806acafab7b', //	Mestská časť Dúbravka
+  '63d264f2965fb806acb00b4d', //	Mestská časť Jarovce
+  '63d28c66965fb806accb5a84', //	Mestská časť Karlova Ves
+  '63d28d17965fb806accbc833', //	Mestská časť Lamač
+  '63d28e23965fb806accc52a5', //	Mestská časť Nové mesto
+  '63d28ed3965fb806accce673', //	Mestská časť Podunajské Biskupice
+  '63d28f93965fb806accd44bf', //	Mestská časť Rača
+  '63d290abbacf0a0f6caa71b9', //	Mestská časť Rusovce
+  '63d29147bacf0a0f6caad199', //	Mestská časť Ružinov
+  '63d29266bacf0a0f6cab8d46', //	Mestská časť Staré mesto
+  '63d29335bacf0a0f6cac04bc', //	Mestská časť Vajnory
+  '63d293c5bacf0a0f6cac3b71', //	Mestská časť Vrakuňa
+  '63d29495bacf0a0f6cac8d68', //	Mestská časť Záhorská Bytrica
+]
