+++ next/components/molecules/OrganizationalStructure/OrganizationalStructure.tsx
@@ -1,6 +1,10 @@
-import { structureFetcher } from '@backend/ms-graph/fetchers/msGraphStructureFetcher'
-import { useTranslations } from 'next-intl'
-import useSWR from 'swr'
+import {
+  getMsGraphStructureQueryKey,
+  msGraphStructureFetcher,
+} from '@backend/ms-graph/fetchers/msGraphStructure.fetcher'
+import { Typography } from '@bratislava/component-library'
+import LoadingSpinner from '@bratislava/ui-bratislava/LoadingSpinner/LoadingSpinner'
+import { useQuery } from '@tanstack/react-query'
 
 import { OrganizationalStructureTopLevelAccordion } from './OrganizationalStructureTopLevelAccordion'
 
@@ -10,19 +14,29 @@ export interface OrganizationalStructureProps {
 
 // TODO add search
 export const OrganizationalStructure = ({ title }: OrganizationalStructureProps) => {
-  const { data } = useSWR('organizationalStructure', structureFetcher)
-  const t = useTranslations()
-  return data ? (
-    <div className="flex flex-col">
-      {/* FIXME Typography. Convert to use Typography. Issue: Header size for not header element */}
-      <div className="text-h3 pb-4">{title}</div>
-      {data.groups.map((group) => (
-        <div key={group.id}>
-          <OrganizationalStructureTopLevelAccordion group={group} />
-        </div>
-      ))}
+  const { data, isLoading, isError, error } = useQuery({
+    queryKey: getMsGraphStructureQueryKey(),
+    queryFn: () => msGraphStructureFetcher(),
+    select: (res) => res.data,
+  })
+
+  if (isLoading) {
+    return <LoadingSpinner />
+  }
+
+  if (isError) {
+    return <div className="whitespace-pre">Error: {JSON.stringify(error, null, 2)}</div>
+  }
+
+  return (
+    <div className="flex flex-col gap-4">
+      <Typography type="h2">{title}</Typography>
+
+      <div className="flex flex-col">
+        {data.groups.map((group) => (
+          <OrganizationalStructureTopLevelAccordion group={group} key={group.id} />
+        ))}
+      </div>
     </div>
-  ) : (
-    <div>{t('loading')}</div>
   )
 }
