+++ next/components/molecules/sections/general/OfficialBoardSection/OfficialBoardSection.tsx
@@ -1,27 +1,52 @@
-import type { ParsedOfficialBoardDocument } from '@backend/services/ginis'
+import {
+  getGinisOfficialBoardQueryKey,
+  ginisOfficialBoardFetcher,
+} from '@backend/ginis/fetchers/ginisOfficialBoard.fetcher'
 import { BasicSearchOld } from '@bratislava/ui-bratislava/BasicSearchOld/BasicSearchOld'
 import { NoResultsFound } from '@bratislava/ui-bratislava/NoResultsFound/NoResultsFound'
+import LoadingSpinner from '@components/ui/LoadingSpinner/LoadingSpinner'
 import { OfficialBoardCards } from '@components/ui/OfficialBoardCards/OfficialBoardCards'
 import { useQuery } from '@tanstack/react-query'
 import { useTranslations } from 'next-intl'
 import React, { useState } from 'react'
 
-const OfficialBoardSection = () => {
-  const [search, setSearch] = useState('')
-
+const DataWrapper = ({ search }: { search: string }) => {
   const t = useTranslations()
 
-  // TODO: Refactor, prefetch and extract fetcher.
-  const { data } = useQuery({
-    queryKey: ['OfficialBoard', search],
-    queryFn: () =>
-      fetch(
-        `/api/ginis/documents-list?${new URLSearchParams({
-          search,
-        })}`,
-      ).then((r) => r.json() as Promise<ParsedOfficialBoardDocument[]>),
+  // TODO prefetch
+  const { data, isLoading, isError, error } = useQuery({
+    queryKey: getGinisOfficialBoardQueryKey(search),
+    queryFn: () => ginisOfficialBoardFetcher(search),
     keepPreviousData: true,
+    select: (res) => res.data,
   })
+
+  if (isLoading) {
+    return <LoadingSpinner />
+  }
+
+  // TODO replace by proper error
+  if (isError) {
+    return <div className="whitespace-pre">Error: {JSON.stringify(error, null, 2)}</div>
+  }
+
+  return data.length > 0 ? (
+    <OfficialBoardCards
+      query={search}
+      title={t('recentlyAddedDocuments')}
+      viewButtonText={t('viewTheDocument')}
+      downloadButtonText={t('download')}
+      documents={data}
+    />
+  ) : (
+    <NoResultsFound title={t('weDidntFindAnything')} message={t('tryEnteringSomethingElse')} />
+  )
+}
+
+const OfficialBoardSection = () => {
+  const t = useTranslations()
+
+  const [search, setSearch] = useState('')
 
   return (
     <>
@@ -33,18 +58,8 @@ const OfficialBoardSection = () => {
         onSubmit={(newSearch) => setSearch(newSearch)}
         initialValue={search}
       />
-      {data && data.length === 0 && (
-        <NoResultsFound title={t('weDidntFindAnything')} message={t('tryEnteringSomethingElse')} />
-      )}
-      {data && data.length > 0 && (
-        <OfficialBoardCards
-          query={search}
-          title={t('recentlyAddedDocuments')}
-          viewButtonText={t('viewTheDocument')}
-          downloadButtonText={t('download')}
-          documents={data}
-        />
-      )}
+
+      <DataWrapper search={search} />
     </>
   )
 }
