--- next/pages/api/facebook-posts.ts
@@ -1,9 +1,0 @@
-import { fetchFacebookPosts } from '@utils/facebook'
-import type { NextApiRequest, NextApiResponse } from 'next'
-
-const handler = async (_req: NextApiRequest, res: NextApiResponse) => {
-  const posts = await fetchFacebookPosts()
-  return res.json(posts)
-}
-
-export default handler
