+++ next/pages/api/ginis/document-load-file/[id].ts
@@ -1,6 +1,6 @@
 // @ts-strict-ignore
+import { ResponseGinisBodyLoadFile } from '@backend/ginis/server/api-data.dto'
 import axios, { AxiosRequestConfig } from 'axios'
-import { ResponseGinisBodyLoadFile } from 'backend/dtos/ginis/api-data.dto'
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { parseString } from 'xml2js'
