+++ next/pages/api/ginis/newest.ts
@@ -1,10 +1,7 @@
-// @ts-strict-ignore
-import {
-  getParsedUDEDocumentsList,
-  mockedParsedDocuments,
-  ParsedOfficialBoardDocument,
-  shouldMockGinis,
-} from 'backend/services/ginis'
+import { mockedParsedDocuments } from '@backend/ginis/mocks'
+import { getParsedUDEDocumentsList } from '@backend/ginis/server/ginisOfficialBoard'
+import { ParsedOfficialBoardDocument } from '@backend/ginis/types'
+import { shouldMockGinis } from '@backend/ginis/utils'
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 // returns the 3 newest documents - used for front page
@@ -12,7 +9,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   let latestOfficialBoard: ParsedOfficialBoardDocument[] = []
   latestOfficialBoard = shouldMockGinis()
     ? mockedParsedDocuments
-    : await getParsedUDEDocumentsList(undefined, 3)
+    : await getParsedUDEDocumentsList()
   return res.json(latestOfficialBoard)
 }
