+++ next/utils/prefetchPageSections.ts
@@ -1,9 +1,17 @@
+import {
+  getGinisOfficialBoardQueryKey,
+  ginisOfficialBoardFetcher,
+} from '@backend/ginis/fetchers/ginisOfficialBoard.fetcher'
 import { PageEntityFragment } from '@backend/graphql'
 import {
   getRelatedBlogPostsQueryKey,
   relatedBlogPostsFetcher,
 } from '@backend/graphql/fetchers/relatedBlogPosts.fetcher'
 import { client } from '@backend/graphql/gql'
+import {
+  getMsGraphStructureQueryKey,
+  msGraphStructureFetcher,
+} from '@backend/ms-graph/fetchers/msGraphStructure.fetcher'
 import { dehydrate, QueryClient } from '@tanstack/react-query'
 
 export const prefetchPageSections = async (page: PageEntityFragment, locale: string) => {
@@ -27,6 +35,16 @@ export const prefetchPageSections = async (page: PageEntityFragment, locale: str
 
   if (sectionTypes.includes('ComponentSectionsInbaArticlesList')) {
     await queryClient.prefetchQuery(['InbaTags', locale], () => client.InbaTags({ locale }))
+  }
+
+  if (sectionTypes.includes('ComponentSectionsOfficialBoard')) {
+    await queryClient.prefetchQuery(getGinisOfficialBoardQueryKey(''), () =>
+      ginisOfficialBoardFetcher(''),
+    )
+  }
+
+  if (sectionTypes.includes('ComponentSectionsOrganizationalStructure')) {
+    await queryClient.prefetchQuery(getMsGraphStructureQueryKey(), () => msGraphStructureFetcher())
   }
 
   await queryClient.prefetchQuery(getRelatedBlogPostsQueryKey(page, locale), () =>
